Bump gradle and kotlin dependencies
Updates the Android Gradle Plugin from 8.0.2 to 8.2.0. This newer version of the AGP offers benefits like improved build speeds and enhanced dependency management.

The Kotlin version is updated from 1.9.10 to 1.9.22. This is a stability and bug-fix release that improves the developer experience.

The Jetpack Compose Compiler extension version is updated from 1.5.3 to 1.5.8 to ensure compatibility with Kotlin 1.9.22.